% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_type.R
\name{cast_type}
\alias{cast_type}
\title{Cast a value to the most appropriate type}
\usage{
cast_type(value)
}
\arguments{
\item{value}{The value to cast.}
}
\value{
The value casted to the most appropriate type.
}
\description{
Cast a value to the most appropriate type
}
