% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_status.R
\name{get_cache_status}
\alias{get_cache_status}
\title{Get the cache status of an endpoint.}
\usage{
get_cache_status(endpoint = NULL)
}
\arguments{
\item{endpoint}{The endpoint to check the cache status of.}
}
\value{
A list with the cache status.
}
\description{
Get the cache status of an endpoint.
}
