% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_endpoint.R
\name{make_endpoint}
\alias{make_endpoint}
\title{Make an endpoint.}
\usage{
make_endpoint(...)
}
\arguments{
\item{...}{The parts of the endpoint.}
}
\value{
The endpoint.
}
\description{
Make an endpoint.
}
