% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_api.R
\name{process_api}
\alias{process_api}
\title{Process an API request.}
\usage{
process_api(endpoint = NULL, cached = FALSE)
}
\arguments{
\item{endpoint}{The endpoint to process.}

\item{cached}{Whether to use the cache or not.}
}
\value{
The response from the API.
}
\description{
Process an API request.
}
