% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_api_with_cache.R
\name{process_api_with_cache}
\alias{process_api_with_cache}
\title{Process an API request with cache.}
\usage{
process_api_with_cache(endpoint, force_refresh = FALSE)
}
\arguments{
\item{endpoint}{The endpoint to process.}

\item{force_refresh}{Whether to force a refresh of the cache.}
}
\value{
A list with the response and the status of the cache.
}
\description{
Process an API request with cache.
}
