% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_safely.R
\name{unlist_safely}
\alias{unlist_safely}
\title{Unlist a list safely}
\usage{
unlist_safely(x)
}
\arguments{
\item{x}{The list to unlist.}
}
\value{
The list unlisted.
}
\description{
Unlist a list safely
}
