% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_api_with_cache.R
\name{process_api_with_cache}
\alias{process_api_with_cache}
\title{Process an API request with cache.}
\usage{
process_api_with_cache(endpoint)
}
\arguments{
\item{endpoint}{The endpoint to process.}
}
\value{
The response from the API.
}
\description{
Process an API request with cache.
}
